+++ src/App.tsx
@@ -1,22 +1,19 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import logo from './assets/images/logo.svg';
+import './styles/App.css';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
         <a
           className="App-link"
-          href="https://reactjs.org"
+          href="https://github.com/yukinoda"
           target="_blank"
           rel="noopener noreferrer"
         >
-          Learn React
+          https://github.com/yukinoda
         </a>
       </header>
     </div>
